requirements.txt
Django==2.0.6
web3==4.3.0
eth_account==0.2.3
requests==2.19.1
djangorestframework==3.8.2
